@@ -4,16 +4,16 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=2.12.1 <3.0.0'
 
 dependencies:
-  camera: ^0.8.0
+  camera: ^0.8.1
   flutter:
     sdk: flutter
   flutter_curiosity:
     path: ../
-  flutter_waya: ^1.7.1
-  permission_handler: ^6.0.1
+  flutter_waya: ^1.10.1
+  permission_handler: ^7.1.0
 
 flutter:
   uses-material-design: true
