@@ -1,5 +1,5 @@
 import 'package:curiosity/main.dart';
-import 'package:curiosity/src/camera/scan.dart';
+import 'package:curiosity/src/camera/scanner.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_curiosity/flutter_curiosity.dart';
 import 'package:flutter_waya/flutter_waya.dart';
@@ -17,23 +17,16 @@ class _CameraGalleryPageState extends State<CameraGalleryPage> {
   @override
   Widget build(BuildContext context) {
     return OverlayScaffold(
-        appBar: AppBar(title: const Text('Camera and Gallery')),
+        appBar: const AppBarText('Camera and Gallery'),
         body: Universal(isScroll: true, children: <Widget>[
-          ElevatedButton(
-              onPressed: () => scan(context), child: const Text('扫码')),
-          ElevatedButton(
-              onPressed: () => systemGallery(), child: const Text('打开系统相册')),
-          ElevatedButton(
-              onPressed: () => systemCamera(), child: const Text('打开系统相机')),
-          ElevatedButton(
-              onPressed: () => scanImage(context),
-              child: const Text('相机识别二维码')),
-          ElevatedButton(
-              onPressed: () => push(UrlImageScanPage()),
-              child: const Text('识别Url二维码')),
-          ElevatedButton(
-              onPressed: () => push(FileImageScanPage()),
-              child: const Text('识别本地图片二维码')),
+          ElevatedText(onPressed: scan, text: '扫码'),
+          ElevatedText(onPressed: systemGallery, text: '打开系统相册'),
+          ElevatedText(onPressed: systemCamera, text: '打开系统相机'),
+          ElevatedText(onPressed: scanImage, text: '相机识别二维码'),
+          ElevatedText(
+              onPressed: () => push(UrlImageScanPage()), text: '识别Url二维码'),
+          ElevatedText(
+              onPressed: () => push(FileImageScanPage()), text: '识别本地图片二维码'),
           const SizedBox(height: 20),
           Container(
               padding: const EdgeInsets.only(top: 100),
@@ -45,7 +38,7 @@ class _CameraGalleryPageState extends State<CameraGalleryPage> {
         ]));
   }
 
-  Future<void> scanImage(BuildContext context) async {
+  Future<void> scanImage() async {
     if (!isMobile) return;
     final bool permission = await requestPermissions(Permission.camera, '相机') &&
         await requestPermissions(Permission.storage, '手机存储');
@@ -56,7 +49,7 @@ class _CameraGalleryPageState extends State<CameraGalleryPage> {
     }
   }
 
-  Future<void> scan(BuildContext context) async {
+  Future<void> scan() async {
     if (!isMobile) return;
     final bool permission = await requestPermissions(Permission.camera, '相机') &&
         await requestPermissions(Permission.storage, '手机存储');
@@ -72,18 +65,18 @@ class _CameraGalleryPageState extends State<CameraGalleryPage> {
   }
 
   Future<void> systemGallery() async {
-    final String data = await openSystemGallery;
+    final String? data = await openSystemGallery();
     showToast(data.toString());
-    text = data ?? '';
+    text = data.toString();
     setState(() {});
   }
 
   Future<void> systemCamera() async {
     if (!isMobile) return;
     if (await requestPermissions(Permission.camera, '使用相机')) {
-      final String data = await openSystemCamera();
+      final String? data = await openSystemCamera();
       showToast(data.toString());
-      text = data ?? '';
+      text = data.toString();
       setState(() {});
     } else {
       showToast('未获取相机权限');
