@@ -14,28 +14,30 @@ class _GetInfoPageState extends State<GetInfoPage> {
   @override
   Widget build(BuildContext context) {
     return OverlayScaffold(
-        appBar: AppBar(title: const Text('App and Device')),
-        body: Universal(isScroll: true, children: <Widget>[
-          ElevatedButton(
-              onPressed: () => getAppInfo(), child: const Text('获取app信息')),
-          ElevatedButton(
-              onPressed: () => getGPS(), child: const Text('获取gps状态')),
-          ElevatedButton(
-              onPressed: () => getDeviceInfo(), child: const Text('获取设备信息')),
-          ElevatedButton(
-              onPressed: () => getInstalled(),
-              child: const Text('获取Android已安装应用')),
-          Column(
-              mainAxisSize: MainAxisSize.min,
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: list),
-        ]));
+      appBar: const AppBarText('App and Device'),
+      body: ScrollList.builder(
+          header: SliverToBoxAdapter(
+              child: Column(children: <Widget>[
+            ElevatedText(onPressed: () => getAppInfo(), text: '获取app信息'),
+            ElevatedText(onPressed: () => getGPS(), text: '获取gps状态'),
+            ElevatedText(onPressed: () => getDeviceInfo(), text: '获取设备信息'),
+            if (isAndroid)
+              ElevatedText(
+                  onPressed: () => getInstalled(), text: '获取Android已安装应用'),
+          ])),
+          placeholder: Container(
+              alignment: Alignment.center,
+              child: const Text('暂无数据'),
+              margin: const EdgeInsets.symmetric(vertical: 30)),
+          itemCount: list.length,
+          itemBuilder: (_, int index) => list[index]),
+    );
   }
 
   Future<void> getInstalled() async {
     final List<AppsModel> data = await getInstalledApp();
     list = <Widget>[];
-    data?.builder((AppsModel appsModel) {
+    data.builder((AppsModel appsModel) {
       final Map<String, dynamic> appModel = appsModel.toJson();
       final List<Widget> app = <Widget>[];
       appModel.forEach((String key, dynamic value) {
@@ -58,9 +60,15 @@ class _GetInfoPageState extends State<GetInfoPage> {
   Future<void> getDeviceInfo() async {
     list = <Widget>[];
     Map<String, dynamic> map = <String, dynamic>{};
-    if (isAndroid) map = (await getAndroidDeviceInfo())?.toJson();
-    if (isIOS) map = (await getIOSDeviceInfo())?.toJson();
-    map?.forEach((String key, dynamic value) {
+    if (isAndroid) {
+      final AndroidDeviceModel? model = await getAndroidDeviceInfo();
+      if (model != null) map = model.toJson();
+    }
+    if (isIOS) {
+      final IOSDeviceModel? model = await getIOSDeviceInfo();
+      if (model != null) map = model.toJson();
+    }
+    map.forEach((String key, dynamic value) {
       if (value is Map) {
         list.add(showText('=== uts', '==='));
         value.forEach((dynamic k, dynamic v) {
@@ -75,7 +83,7 @@ class _GetInfoPageState extends State<GetInfoPage> {
   }
 
   Future<void> getAppInfo() async {
-    final AppInfoModel data = await getPackageInfo();
+    final AppInfoModel? data = await getPackageInfo();
     if (data == null) return;
     final Map<String, dynamic> map = data.toJson();
     list = <Widget>[];
@@ -86,7 +94,7 @@ class _GetInfoPageState extends State<GetInfoPage> {
   }
 
   Future<void> getGPS() async {
-    final bool data = await getGPSStatus;
+    final bool data = await getGPSStatus();
     showToast(data.toString());
   }
 }
