@@ -1,3 +1,4 @@
+import 'package:curiosity/main.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_curiosity/flutter_curiosity.dart';
 import 'package:flutter_waya/flutter_waya.dart';
@@ -13,20 +14,15 @@ class _SharePageState extends State<SharePage> {
   @override
   Widget build(BuildContext context) {
     return OverlayScaffold(
-        appBar: AppBar(title: const Text('Share')),
+        appBar: const AppBarText('Share'),
         body: Universal(
             isScroll: true,
             mainAxisAlignment: MainAxisAlignment.center,
             children: <Widget>[
-              ElevatedButton(
-                  onPressed: () => systemGallery(),
-                  child: const Text('打开系统相册')),
-              ElevatedButton(
-                  onPressed: () => shareText(), child: const Text('分享文字')),
-              ElevatedButton(
-                  onPressed: () => shareImage(), child: const Text('分享图片')),
-              ElevatedButton(
-                  onPressed: () => shareImages(), child: const Text('分享多张图片')),
+              ElevatedText(onPressed: () => systemGallery(), text: '打开系统相册'),
+              ElevatedText(onPressed: shareText, text: '分享文字'),
+              ElevatedText(onPressed: () => shareImage(), text: '分享图片'),
+              ElevatedText(onPressed: () => shareImages(), text: '分享多张图片'),
               const SizedBox(height: 20),
               Column(
                   mainAxisSize: MainAxisSize.min,
@@ -35,13 +31,13 @@ class _SharePageState extends State<SharePage> {
   }
 
   Future<void> systemGallery() async {
-    final String data = await openSystemGallery;
+    final String? data = await openSystemGallery();
     list.add(data.toString());
     setState(() {});
   }
 
   void shareText() {
-    systemShare(title: '分享图片', content: '分享几个文字', shareType: ShareType.text);
+    openSystemShare(title: '分享图片', content: '分享几个文字', shareType: ShareType.text);
   }
 
   void shareImage() {
@@ -49,7 +45,7 @@ class _SharePageState extends State<SharePage> {
       showToast('请先选择图片');
       return;
     }
-    systemShare(title: '分享图片', content: list[0], shareType: ShareType.image);
+    openSystemShare(title: '分享图片', content: list[0], shareType: ShareType.image);
   }
 
   void shareImages() {
@@ -59,7 +55,7 @@ class _SharePageState extends State<SharePage> {
     }
     final List<String> listPath = <String>[];
     listPath.addAll(list);
-    systemShare(
+    openSystemShare(
         title: '分享图片', imagesPath: listPath, shareType: ShareType.images);
   }
 }
