@@ -3,9 +3,6 @@
 
 @interface NativeTools : NSObject
 
-//跳转拨号
-+ (BOOL)callPhone:(NSString *)phoneNumber;
-
 //获取app信息
 + (NSDictionary *)getAppInfo;
 
@@ -13,16 +10,18 @@
 + (NSDictionary *)getDeviceInfo;
 
 //调用系统分享
-+ (void)systemShare:(FlutterMethodCall*)call result:(FlutterResult)result;
-
++ (void)openSystemShare:(FlutterMethodCall*)call result:(FlutterResult)result;
 
 //跳转到APP权限设置页面
-+ (BOOL)jumpAppSetting;
++ (BOOL)openAppSetting;
 
 //判断GPS是否开启，GPS或者AGPS开启一个就认为是开启的
 + (BOOL)getGPSStatus;
 
-// 获取文件或文件夹大小
-+ (NSString *)getFilePathSize:(NSString *)props;
+//能否打开url
++ (BOOL) canOpenURL:(NSString *)url;
+//打开url
++ (void) openURL:(NSDictionary *)arguments :(FlutterResult)result ;
+
 
 @end
