@@ -4,17 +4,6 @@
 
 @implementation NativeTools
 
-+ (BOOL)callPhone:(NSString *)phoneNumber {
-    NSURL *url=[NSURL URLWithString:[@"tel:"stringByAppendingString:phoneNumber]];
-    if([[NSWorkspace sharedWorkspace] URLForApplicationToOpenURL:url]){
-        [[NSWorkspace sharedWorkspace] openURL:url];
-        return YES;
-    }else{
-        return NO;
-    }
-}
-
-
 + (NSDictionary *)getAppInfo{
     NSMutableDictionary *info = [NSMutableDictionary dictionary];
     NSDictionary *app = [[NSBundle mainBundle] infoDictionary];
@@ -35,55 +24,6 @@
 }
 
 
-//获取目录文件或文件夹大小
-+ (NSString *)getFilePathSize:(NSString *)path{
-    // 获取“path”文件夹下的所有文件
-    NSArray *subPathArr = [[NSFileManager defaultManager] subpathsAtPath:path];
-    NSString *filePath  = nil;
-    NSInteger totalSize = 0;
-    for (NSString *subPath in subPathArr){
-        // 1. 拼接每一个文件的全路径
-        filePath =[path stringByAppendingPathComponent:subPath];
-        // 2. 是否是文件夹，默认不是
-        BOOL isDirectory = [Tools isDirectory:path];
-        // 3. 判断文件是否存在
-        BOOL isExist = [Tools isDirectoryExist:path];
-        // 4. 以上判断目的是忽略不需要计算的文件
-        if (@available(macOS 10.10, *)) {
-            if (!isExist || isDirectory || [filePath containsString:@".DS"]){
-                // 过滤: 1. 文件夹不存在  2. 过滤文件夹  3. 隐藏文件
-                continue;
-            }
-        }
-        // 5. 指定路径，获取这个路径的属性
-        NSDictionary *dict = [[NSFileManager defaultManager] attributesOfItemAtPath:filePath error:nil];
-        /**
-         attributesOfItemAtPath: 文件夹路径
-         该方法只能获取文件的属性, 无法获取文件夹属性, 所以也是需要遍历文件夹的每一个文件的原因
-         */
-        // 6. 获取每一个文件的大小
-        if (@available(macOS 10.8, *)) {
-            NSInteger size = [dict[@"NSFileSize"] integerValue];
-            // 7. 计算总大小
-            totalSize += size;
-        } else {
-            // Fallback on earlier versions
-        }
-    }
-    //8. 将文件夹大小转换为 M/KB/B
-    NSString *totalStr = nil;
-    if (totalSize > 1000 * 1000){
-        totalStr = [NSString stringWithFormat:@"%.2fMB",totalSize / 1000.00f /1000.00f];
-        
-    }else if (totalSize > 1000){
-        totalStr = [NSString stringWithFormat:@"%.2fKB",totalSize / 1000.00f ];
-        
-    }else{
-        totalStr = [NSString stringWithFormat:@"%.2fB",totalSize / 1.00f];
-    }
-    return totalStr;
-}
-
 //判断GPS是否开启，GPS或者AGPS开启一个就认为是开启的
 + (BOOL) getGPSStatus {
     if (@available(macOS 10.7, *)) {
@@ -91,5 +31,19 @@
     }
     return  NO;
 }
+
+//能否打开url
++ (BOOL) canOpenUrl:(NSString *)url {
+    NSURL *nsUrl = [[NSURL alloc] initWithString:url];
+    return [[NSWorkspace sharedWorkspace] URLForApplicationToOpenURL:nsUrl] != nil;
+}
+//打开url
++ (void) openUrl:(NSDictionary *)arguments {
+    NSURL *nsUrl = [[NSURL alloc] initWithString:arguments[@"url"]];
+    [[NSWorkspace sharedWorkspace] openURL:nsUrl];
+}
+
+
+
 
 @end
