@@ -8,15 +8,11 @@
     NSLog(@"CuriosityLog = %@", info);
 }
 
-+ (NSString *)resultInfo:(NSString *)info{
-    return [NSString stringWithFormat:@"Curiosity:%@",info];
-}
-
 + (NSString *)resultFail{
-    return @"Curiosity:fail";
+    return @"fail";
 }
 + (NSString *)resultSuccess{
-    return @"Curiosity:success";
+    return @"success";
 }
 // 沙盒是否有指定路径文件夹或文件
 +(BOOL)isDirectoryExist:(NSString *)path {
